refactor: Use u64 for the timeout
BREAKING CHANGE: This changes the type of the function signature.
                 If you used u32 before, simply add `as u64` where
                 passing the value.

+++ src/lib.rs
@@ -47,8 +47,8 @@ impl<T> TimedStack<T> {
         self.len() == 0
     }
 
-    pub fn pop(&self, timeout_ms: u32) -> Option<T> {
-        let timeout_ms = Duration::from_millis(timeout_ms as u64);
+    pub fn pop(&self, timeout_ms: u64) -> Option<T> {
+        let timeout_ms = Duration::from_millis(timeout_ms);
         let mut queue = self.queue.lock().unwrap();
 
         loop {
